fix presenter
